Move loadImage() for TextView to SpeechSessionItem
